PAK-71: Cart Page Done

+++ database/migrations/2023_04_12_201856_create_products_table.php
@@ -25,17 +25,17 @@ return new class extends Migration
             $table->string('permalink', 200)->nullable()->unique();
             $table->string('sku', 50)->nullable()->unique();
 
-            $table->float('price')->default(0);
-            $table->float('discounted_price')->default(0);
+            $table->decimal('price', 8, 2)->default(0);
+            $table->decimal('discounted_price', 8, 2)->default(0);
 
             $table->boolean('call_for_final_rates')->default(false);
             $table->boolean('are_rates_printed_on_package')->default(false);
 
-            $table->float('length')->default(0);
-            $table->float('width')->default(0);
-            $table->float('height')->default(0);
+            $table->decimal('length', 8, 2)->default(0);
+            $table->decimal('width', 8, 2)->default(0);
+            $table->decimal('height', 8, 2)->default(0);
 
-            $table->float('weight')->default(0);
+            $table->decimal('weight', 8, 2)->default(0);
 
             $table->text('short_description')->nullable();
             $table->longText('long_description')->nullable();
